test(s3select): keep fixture within existing s3s boundary
